refactor(core): retire the AgentRunHeader as a record of the run
The header was a third authority over facts the event spine already owns:
route, configuration, root, lineage and terminal status. Every reconciliation
path in the codebase exists because those three could disagree.

The opening RuntimeEvent is now the only record of how an invocation was
opened, and the terminal RuntimeEvent the only record of how it ended.
`@maka/core/runtime-invocation` owns the concept; `AgentRunStore` keeps only
the AgentRunEvent operational ledger, and its `core_agent_runs` row keeps only
what events hang off. The header's decoder moves to `@maka/storage` beside the
migration that consumes it, which is what stops it becoming a live authority
again.

Generated-by: Claude Code
