+++ packages/core/package.json
@@ -30,6 +30,8 @@
     "./collaboration": "./dist/collaboration.js",
     "./orchestration": "./dist/orchestration.js",
     "./tool-mode": "./dist/tool-mode.js",
+    "./record-schema": "./dist/record-schema.js",
+    "./runtime-invocation": "./dist/runtime-invocation.js",
     "./plan": "./dist/plan.js",
     "./agent-run": "./dist/agent-run.js",
     "./subagent-workspace": "./dist/subagent-workspace.js",
