+++ packages/core/src/backend-types.ts
@@ -42,7 +42,7 @@ import type { InteractionClosureReason, InteractionFormResult } from './interact
 import type { RuntimeEvent } from './runtime-event.js';
 import type { SandboxBoundaryResponse, SandboxBoundarySettlement } from './sandbox-boundary.js';
 import type { StoredMessage, PersistedBackendKind } from './session.js';
-import type { AgentRunHeader } from './agent-run.js';
+import type { RuntimeInvocationRecord } from './runtime-invocation.js';
 import type { UserQuestionResponse } from './user-question.js';
 import type { ContextBudgetDiagnostic } from './usage-stats/types.js';
 import type { EffectiveOrchestration } from './orchestration.js';
@@ -94,11 +94,11 @@ export interface BackendSendInput {
    */
   runtimeContext?: RuntimeEvent[];
   /**
-   * Existing durable run headers for `runtimeContext`, used only to verify
+   * The invocations `runtimeContext` came from, used only to verify
    * provider-owned replay against the current model route. RuntimeEvents stay
-   * the transcript authority; route provenance remains owned by AgentRun.
+   * the transcript authority; route provenance is read off each opening fact.
    */
-  runtimeContextRunHeaders?: readonly AgentRunHeader[];
+  runtimeContextInvocations?: readonly RuntimeInvocationRecord[];
   /** Continue from an already committed RuntimeEvent boundary without adding another user turn. */
   continuation?: RuntimeContinuationMetadata;
   /**
@@ -189,8 +189,8 @@ export interface BackendCompactHistoryInput {
    */
   runId: string;
   runtimeContext: readonly RuntimeEvent[];
-  /** Source-run route authority for provider-owned history projected into the compaction call. */
-  runtimeContextRunHeaders?: readonly AgentRunHeader[];
+  /** Source-invocation route authority for provider-owned history projected into the compaction call. */
+  runtimeContextInvocations?: readonly RuntimeInvocationRecord[];
 }
 
 export interface BackendCompactHistoryResult {
