+++ packages/core/src/execution-inspect.ts
@@ -17,7 +17,6 @@
  * under the License.
  */
 
-import { AGENT_RUN_STATUSES, type AgentRunHeader } from './agent-run.js';
 import { EXECUTION_LOG_LEDGERS, type ExecutionLogCoverage } from './execution-log-coverage.js';
 import { SESSION_STATUSES, type SessionHeader } from './session.js';
 
@@ -36,20 +35,22 @@ export interface ExecutionInspectDiagnostic {
   eventId?: string;
 }
 
+export const AGENT_RUN_INSPECT_STATUSES = ['running', 'completed', 'failed', 'cancelled'] as const;
+
 export interface AgentRunInspectIdentity {
   sessionId: string;
   agentRunId: string;
-  invocationId?: string;
+  invocationId: string;
   turnId: string;
   parentRunId?: string;
   resumedFromRunId?: string;
   retriedFromRunId?: string;
   parentTurnId?: string;
   agentId?: string;
-  status: AgentRunHeader['status'];
-  createdAt: number;
-  updatedAt: number;
-  completedAt?: number;
+  /** Derived from the terminal RuntimeEvent; `running` means there is none yet. */
+  status: (typeof AGENT_RUN_INSPECT_STATUSES)[number];
+  openedAt: number;
+  endedAt?: number;
   failureClass?: string;
   abortSource?: string;
 }
@@ -80,8 +81,6 @@ export interface AgentRunInspectCompactionCheckpoint {
 export interface AgentRunInspectSourceHealth {
   runtimeLedger: 'present' | 'missing' | 'read_failed';
   runtimeTerminalPresent: boolean;
-  operationalTerminalPresent: boolean;
-  statusConsistency: 'consistent' | 'inconsistent' | 'incomplete';
 }
 
 export interface AgentRunInspectDocument {
@@ -194,28 +193,28 @@ function isAgentRunIdentity(value: unknown): value is AgentRunInspectIdentity {
   return (
     hasShape(
       value,
-      ['sessionId', 'agentRunId', 'turnId', 'status', 'createdAt', 'updatedAt'],
+      ['sessionId', 'agentRunId', 'invocationId', 'turnId', 'status', 'openedAt'],
       [
-        'invocationId',
         'parentRunId',
         'resumedFromRunId',
         'retriedFromRunId',
         'parentTurnId',
         'agentId',
-        'completedAt',
+        'endedAt',
         'failureClass',
         'abortSource',
       ],
     ) &&
     isString(value.sessionId) &&
     isString(value.agentRunId) &&
+    isString(value.invocationId) &&
     isString(value.turnId) &&
-    AGENT_RUN_STATUSES.includes(value.status as (typeof AGENT_RUN_STATUSES)[number]) &&
-    isCount(value.createdAt) &&
-    isCount(value.updatedAt) &&
-    isOptionalCount(value.completedAt) &&
+    AGENT_RUN_INSPECT_STATUSES.includes(
+      value.status as (typeof AGENT_RUN_INSPECT_STATUSES)[number],
+    ) &&
+    isCount(value.openedAt) &&
+    isOptionalCount(value.endedAt) &&
     [
-      value.invocationId,
       value.parentRunId,
       value.resumedFromRunId,
       value.retriedFromRunId,
@@ -237,24 +236,11 @@ function isAgentRunSources(value: unknown): boolean {
     isCount(value.operationalEventCount) &&
     isCount(value.runtimeEventCount) &&
     (value.runtimeCoverage === undefined || isCoverage(value.runtimeCoverage)) &&
-    hasShape(
-      value.health,
-      [
-        'runtimeLedger',
-        'runtimeTerminalPresent',
-        'operationalTerminalPresent',
-        'statusConsistency',
-      ],
-      [],
-    ) &&
+    hasShape(value.health, ['runtimeLedger', 'runtimeTerminalPresent'], []) &&
     (value.health.runtimeLedger === 'present' ||
       value.health.runtimeLedger === 'missing' ||
       value.health.runtimeLedger === 'read_failed') &&
-    typeof value.health.runtimeTerminalPresent === 'boolean' &&
-    typeof value.health.operationalTerminalPresent === 'boolean' &&
-    (value.health.statusConsistency === 'consistent' ||
-      value.health.statusConsistency === 'inconsistent' ||
-      value.health.statusConsistency === 'incomplete')
+    typeof value.health.runtimeTerminalPresent === 'boolean'
   );
 }
