+++ packages/storage/src/__tests__/usage-stores.test.ts
@@ -735,8 +735,8 @@ function appendModelCallAuthorityEvent(
     lease.transaction('write', () => {
       lease.database
         .prepare(`
-          INSERT INTO core_agent_runs(session_id, run_id, created_at, record_json)
-          VALUES (?, ?, 0, '{}')
+          INSERT INTO core_agent_runs(session_id, run_id, created_at)
+          VALUES (?, ?, 0)
         `)
         .run(value.sessionId, value.runId);
       lease.database
