+++ packages/storage/src/__tests__/workhub-coordination-root-admission.test.ts
@@ -22,7 +22,7 @@ import { mkdtemp, rm } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { test } from 'node:test';
-import type { RootExecutionDescriptor } from '@maka/core/agent-run';
+import type { RootExecutionDescriptor } from '@maka/core/runtime-invocation';
 import { createSqliteAgentRunStore } from '../agent-run-store.js';
 
 test('WorkHub Coordination admission preserves its bounded content identity across restart', async () => {
